@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -14,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ApiTypeModel {
 
-    public static final ApiTypeModel UNKNOWN = new ApiTypeModel("unknown", ApiTypeType.UNKNOWN);
+    public static final ApiTypeModel UNKNOWN = new ApiTypeModel("unknown", ApiTypeType.UNKNOWN, null);
 
     private String name;
     private ApiTypeType type;
@@ -22,17 +21,14 @@ public class ApiTypeModel {
     private boolean required;
     private List<ApiTypeModel> typeArguments;
     private List<String> nesting;
-
-    public ApiTypeModel(String name, ApiTypeType type) {
-        this(name, type, null);
-    }
+    private List<String> annotations;
 
     public ApiTypeModel(String name, ApiTypeType type, boolean required) {
         this(name, type, null, required);
     }
 
     public ApiTypeModel(String name, ApiTypeType type, boolean required, List<ApiTypeModel> typeArguments) {
-        this(name, type, null, required, typeArguments, Collections.emptyList());
+        this(name, type, null, required, typeArguments, Collections.emptyList(), Collections.emptyList());
     }
 
     public ApiTypeModel(String name, ApiTypeType type, String className) {
@@ -40,6 +36,6 @@ public class ApiTypeModel {
     }
 
     public ApiTypeModel(String name, ApiTypeType type, String className, boolean required) {
-        this(name, type, className, required, Collections.emptyList(), Collections.emptyList());
+        this(name, type, className, required, Collections.emptyList(), Collections.emptyList(), Collections.emptyList());
     }
 }
